requirements.txt
discord.py~=2.3.2
PyQt5~=5.15.10
requests~=2.31.0
python-dotenv~=1.0.1
urlextract~=1.9.0
youtube-dl~=2021.12.17
#pip install PyQt5 PyQtChart
humanize~=4.9.0